requirements.txt
requests==2.31.0
urllib3==2.0.7
schedule==1.2.0
python-dateutil==2.8.2
